MatchesList and Contacts Fixes: - Remove unknown user text in front of the number, show only number - User has to explicitly click on ‘Import Contacts’
