+++ src/components/ContactsList.tsx
@@ -5,14 +5,14 @@ import { List, ListItem } from 'material-ui/List';
 import Divider from 'material-ui/Divider';
 import Subheader from 'material-ui/Subheader';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
-// import Snackbar from 'material-ui/Snackbar';
 import RaisedButton from 'material-ui/RaisedButton';
 import ContentAdd from 'material-ui/svg-icons/content/add';
 import AutoComplete from 'material-ui/AutoComplete';
 import MenuItem from 'material-ui/MenuItem';
 import { ourFirebase } from '../services/firebase';
 import { connect } from 'react-redux';
-import { StoreState } from '../types/index';
+import { StoreState, PhoneNumberToContact } from '../types/index';
+import { store } from '../stores/index';
 import { History } from 'history';
 import {
   checkNotNull,
@@ -20,7 +20,8 @@ import {
   isIos,
   findMatch,
   getPhoneNumberToUserInfo,
-  checkPhoneNumber
+  checkPhoneNumber,
+  UNKNOWN_NAME
 } from '../globals';
 
 const style: React.CSSProperties = {
@@ -60,8 +61,9 @@ interface Props {
   searchByNumber: boolean;
 }
 
+declare let ContactFindOptions: any;
+
 class ContactsList extends React.Component<Props, {}> {
-  // timer: any = undefined;
 
   // TODO: if Object.keys(state.phoneNumberToContact)=[] (i.e., the user didn't give
   // the permission to fetch contacts), then let's ALSO ask the user to enter his
@@ -74,7 +76,78 @@ class ContactsList extends React.Component<Props, {}> {
   state = {
     filterValue: '',
     message: 'Message sent',
-    notUsers: this.props.notUsers
+    notUsers: this.props.notUsers,
+    didFetchContacts: false
+  };
+
+  fetchContacts = () => {
+    // TODO: show per-premission screen.
+    if (!navigator.contacts) {
+      console.error('No navigator.contacts!');
+      return;
+    }
+    console.log('Fetching contacts');
+
+    var options = new ContactFindOptions();
+    options.filter = '';
+    options.multiple = true;
+    options.desiredFields = [
+      navigator.contacts.fieldType.displayName,
+      navigator.contacts.fieldType.phoneNumbers
+    ];
+    options.hasPhoneNumber = true;
+    navigator.contacts.find(['*'], this.onSuccess, this.onError, options);
+  };
+
+  onSuccess = (contacts: any[]) => {
+    console.log('Successfully got contacts: ', contacts);
+    let myCountryCode = store.getState().myUser.myCountryCode;
+    if (!myCountryCode) {
+      console.error('Missing country code');
+      return;
+    }
+    if (!contacts) {
+      console.error('Missing contacts');
+      return;
+    }
+    let currentContacts: PhoneNumberToContact = {};
+    for (let contact of contacts) {
+      if (!contact.phoneNumbers) {
+        continue;
+      }
+      for (let phoneNumber of contact.phoneNumbers) {
+        const localNumber = phoneNumber['value'].replace(/[^0-9]/g, '');
+        const phoneInfo: PhoneNumInfo | null = checkPhoneNumber(
+          localNumber,
+          myCountryCode
+        );
+        if (
+          phoneInfo &&
+          phoneInfo.isPossibleNumber &&
+          phoneInfo.isValidNumber &&
+          phoneInfo.maybeMobileNumber
+        ) {
+          const internationalNumber = phoneInfo.e164Format;
+          if (ourFirebase.checkPhoneNum(internationalNumber)) {
+            console.error(
+              'e164Format returned illegal phone number:',
+              internationalNumber
+            );
+            continue;
+          }
+          const newContact: Contact = {
+            name: contact.displayName,
+            phoneNumber: internationalNumber
+          };
+          currentContacts[internationalNumber] = newContact;
+        }
+      }
+    }
+    ourFirebase.storeContacts(currentContacts);
+  };
+
+  onError = () => {
+    console.error('Error fetching contacts');
   };
 
   handleRequest = (chosenRequest: DataSourceConfig, index: number) => {
@@ -250,7 +323,16 @@ class ContactsList extends React.Component<Props, {}> {
     return (
       <div>
         <div style={searchStyle}>{searchField}</div>
-
+        {this.state.didFetchContacts ? null : <RaisedButton 
+          onClick={() => {
+            this.fetchContacts()
+            this.setState({
+              didFetchContacts: true
+            })
+          }}
+          label={'Import My Contacts'}
+          style={searchStyle}
+        />}
         <List>
           <Subheader>Game User</Subheader>
           {this.filterParticipants(this.filterUsers(this.props.users)).map(
@@ -283,7 +365,7 @@ class ContactsList extends React.Component<Props, {}> {
               <ListItem
                 key={contact.phoneNumber}
                 primaryText={
-                  contact.name +
+                  (contact.name === UNKNOWN_NAME ? '' : contact.name + ' ') +
                   (parsed && parsed.isValidNumber
                     ? `(${parsed.internationalFormat})`
                     : '')
